+++ src/plugins/chatview/templates/chat-head.js
